Move filter to its own component.
